cloud-proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
httpx==0.25.1
pydantic==2.5.0
